Move @EnableTransactionManagement to HibernateConfig Add Activate,Deactivate and isAuthenticated methods and unit tests

+++ src/main/java/com/example/GymProject/service/TraineeService.java
@@ -41,6 +41,7 @@ public class TraineeService {
     @Autowired
     private EntityMapper entityMapper;
 
+    @Transactional
     public TraineeDto createTrainee(TraineeDto traineeDTO) {
         Assert.notNull(traineeDTO, "TraineeDto cannot be null");
         Trainee trainee = entityMapper.toTrainee(traineeDTO);
@@ -53,30 +54,41 @@ public class TraineeService {
         trainee.setUser(user);
         Trainee trainee1 = traineeDao.createTrainee(trainee);
         return entityMapper.toTraineeDto(trainee1);
+
     }
 
     public TraineeDto getTraineeByUsername(String username, String password) {
         Assert.notNull(username, "Username cannot be null");
-        authenticate(username, password);
+        if(isAuthenticated(username, password)){
         Trainee trainee = traineeDao.getTraineeByUsername(username);
         return entityMapper.toTraineeDto(trainee);
+    }
+        logger.error("Authentication failed for trainee {}",username);
+        return null;
+
     }
 
 
     public TraineeDto updateTrainee(TraineeDto traineeDTO, String password) {
         Assert.notNull(traineeDTO, "TraineeDto cannot be null");
-        authenticate(traineeDTO.getUserDto().getUsername(), password);
-        Trainee trainee = entityMapper.toTrainee(traineeDTO);
-        User user = entityMapper.toUser(traineeDTO.getUserDto());
-        trainee.setUser(user);
-        return entityMapper.toTraineeDto(traineeDao.updateTrainee(trainee));
+        if(isAuthenticated(traineeDTO.getUserDto().getUsername(), password)) {
+            Trainee trainee = entityMapper.toTrainee(traineeDTO);
+            User user = entityMapper.toUser(traineeDTO.getUserDto());
+            trainee.setUser(user);
+            return entityMapper.toTraineeDto(traineeDao.updateTrainee(trainee));
+        }
+        logger.error("Authentication failed for trainee {}",traineeDTO.getUserDto().getUsername());
+        return null;
     }
 
-    //pordzel
+
     public void deleteTraineeByUsername(String username, String password) {
         Assert.notNull(username, "Username cannot be null");
-        authenticate(username, password);
-        traineeDao.deleteTraineeByUsername(username);
+        if(isAuthenticated(username, password)) {
+            traineeDao.deleteTraineeByUsername(username);
+        }
+        logger.error("Authentication failed for trainee {}",username);
+
     }
 
     public boolean matchUsernameAndPassword(String username, String password) {
@@ -89,76 +101,103 @@ public class TraineeService {
     public void changeTraineePassword(String username, String newPassword, String password) {
         Assert.notNull(username, "Username cannot be null");
         Assert.notNull(newPassword, "New password cannot be null");
-        authenticate(username, password);
-        UserDto userDTO = userService.getUserByUsername(username);
-        if (userDTO != null) {
-            if (!password.equals(newPassword)) {
-                userDTO.setPassword(newPassword);
-                userService.updateUser(userDTO);
-                logger.info("Password has successfully changed for trainee {}", username);
-            } else {
-                logger.warn("Cannot change password for user {} since new password is equal to old password", username);
+       if(isAuthenticated(username, password)) {
+           UserDto userDTO = userService.getUserByUsername(username);
+           if (userDTO != null) {
+               if (!password.equals(newPassword)) {
+                   userDTO.setPassword(newPassword);
+                   userService.updateUser(userDTO);
+                   logger.info("Password has successfully changed for trainee {}", username);
+               } else {
+                   logger.warn("Cannot change password for user {} since new password is equal to old password", username);
 
-            }
-        } else {
-            logger.warn("Cannot change password for user {} since such user was not found", username);
-        }
+               }
+           } else {
+               logger.warn("Cannot change password for user {} since such user was not found", username);
+           }
+       }
+        logger.error("Authentication failed for trainee {}",username);
+
     }
 
-    public void activateDeactivateTrainee(String username, boolean isActive, String password) {
+
+    public void activate(String username, String password){
         Assert.notNull(username, "Username cannot be null");
-        authenticate(username, password);
-        Trainee trainee = traineeDao.getTraineeByUsername(username);
-        if (trainee != null) {
-            logger.info("Activation/Deactivation was successfully performed for trainee {}", username);
-            trainee.getUser().setIsActive(isActive);
-            userService.updateUser(entityMapper.toUserDto(trainee.getUser()));
+        if(isAuthenticated(username, password)) {
+            Trainee trainee = traineeDao.getTraineeByUsername(username);
+            if (trainee != null) {
+                logger.info("Activation was successfully performed for trainee {}", username);
+                trainee.getUser().setIsActive(true);
+                userService.updateUser(entityMapper.toUserDto(trainee.getUser()));
+            }
         }
+        logger.error("Authentication failed for trainee {}",username);
+
+    }
+    public void deactivate(String username, String password){
+        Assert.notNull(username, "Username cannot be null");
+        if(isAuthenticated(username, password)) {
+            Trainee trainee = traineeDao.getTraineeByUsername(username);
+            if (trainee != null) {
+                logger.info("Deactivation was successfully performed for trainee {}", username);
+                trainee.getUser().setIsActive(true);
+                userService.updateUser(entityMapper.toUserDto(trainee.getUser()));
+            }
+        }
+        logger.error("Authentication failed for trainee {}",username);
+
     }
 
     public List<TrainingDto> getTraineeTrainings(String username, Date fromDate, Date toDate, String trainerName, String trainingType, String password) {
         Assert.notNull(username, "Username cannot be null");
-        authenticate(username, password);
+        if(isAuthenticated(username, password)){
         List<Training> trainings = traineeDao.getTraineeTrainings(username, fromDate, toDate, trainerName, trainingType);
         logger.info("Getting trainings for trainee {}", username);
         return trainings.stream()
                 .map(entityMapper::toTrainingDto)
                 .collect(Collectors.toList());
     }
+        logger.error("Authentication failed for trainee {}",username);
+        return null;
+    }
 
     public void updateTraineeTrainers(String username, Set<TrainerDto> trainerDtos, String password) {
         Assert.notNull(username, "Username cannot be null");
-        authenticate(username, password);
-        Trainee trainee = traineeDao.getTraineeByUsername(username);
-        logger.info("Updating trainers for trainee {}", username);
-        Set<Trainer> trainers = trainerDtos.stream()
-                .map(entityMapper::toTrainer)
-                .collect(Collectors.toSet());
-        trainee.setTrainers(trainers);
-        traineeDao.updateTrainee(trainee);
+        if(isAuthenticated(username, password)) {
+            Trainee trainee = traineeDao.getTraineeByUsername(username);
+            logger.info("Updating trainers for trainee {}", username);
+            Set<Trainer> trainers = trainerDtos.stream()
+                    .map(entityMapper::toTrainer)
+                    .collect(Collectors.toSet());
+            trainee.setTrainers(trainers);
+            traineeDao.updateTrainee(trainee);
+        }
+        logger.error("Authentication failed for trainee {}",username);
+
     }
 
     @Transactional(readOnly = true)
     public List<TrainerDto> getUnassignedTrainers(String traineeUsername, String password) {
-        authenticate(traineeUsername, password);
-        Trainee trainee = traineeDao.getTraineeByUsername(traineeUsername);
-        List<Trainer> allTrainers = trainerDao.getAllTrainers();
+        if(isAuthenticated(traineeUsername, password)) {
+            Trainee trainee = traineeDao.getTraineeByUsername(traineeUsername);
+            List<Trainer> allTrainers = trainerDao.getAllTrainers();
 
-        List<Trainer> assignedTrainers = trainee.getTrainers().stream()
-                .collect(Collectors.toList());
+            List<Trainer> assignedTrainers = trainee.getTrainers().stream()
+                    .collect(Collectors.toList());
 
-        List<Trainer> unassignedTrainers = allTrainers.stream()
-                .filter(trainer -> !assignedTrainers.contains(trainer))
-                .collect(Collectors.toList());
+            List<Trainer> unassignedTrainers = allTrainers.stream()
+                    .filter(trainer -> !assignedTrainers.contains(trainer))
+                    .collect(Collectors.toList());
 
-        return unassignedTrainers.stream()
-                .map(entityMapper::toTrainerDto)
-                .collect(Collectors.toList());
+            return unassignedTrainers.stream()
+                    .map(entityMapper::toTrainerDto)
+                    .collect(Collectors.toList());
+        }
+        logger.error("Authentication failed for trainee {}",traineeUsername);
+        return null;
     }
 
-    public void authenticate(String username, String password) {
-        if (!matchUsernameAndPassword(username, password)) {
-            throw new SecurityException("Authentication failed for username: " + username);
-        }
+    public boolean isAuthenticated(String username,String password){
+        return matchUsernameAndPassword(username,password);
     }
 }
